Move validate_task_schema() to client.py
